Changes i want: in automation window -> automation tab -> custom modules, when creating a custom module, i want a new module type "Label" with simple text input. when it is saved, it is saved as "{the input text}:" (in the code, it should be saved as an item of a dedicated list)
and i want another new module type "Goto" with combobox to select a single label (from the list of labels)

In the session pane, if the "Goto Label_A" is reached, the process moves to the place where "Label_A" label is and go on.
it is just like classic "goto" of the BASIC or C.

i want additional module type "If Message {custom procedure} is Yes" and a built-in control "Else If Message is No".
when "If Message {custom procedure} is Yes" is run, a YesNo messagebox with result of {custom procedure} as text shows. if Yes is clicked, the next procedure is run sequentially and when met "Else If Message is No" module, jump to "end if". if No is clicked, if there is "Else If Message is No" module afterward, immediately jump to that module, if not, jump to "end if".

as this is a "if" module, it should be ended with "end if". and the "Else If Message is No" may or may not exist, but it can only exist between "If Message {custom procedure} is Yes" and "end if". this should be validated when saving automation.