Current state: the central (azure sql) connection string is no longer needed as all the transaction is occured via API. ======================================================== Changes i want: In settings window -> Network, can you delete the central (azure sql) connection string label and textbox and its related codes? ======================================================== Current state: the local/intranet connection string in the settings window -> network tab should include the password, thus the password is exposed to the user. ======================================================== Changes i want: i want password hidden and hard coded as "`123qweas" thus, if as-is connection string value in the textbox was "Host=127.0.0.1;Port=5432;Database=wysg_dev;Username=postgres;Password=`123qweas"
the to-be value should be "Host=127.0.0.1;Port=5432;Database=wysg_dev;Username=postgres" and in the code, the ";Password=`123qweas" should be appended, thus resulting in "Host=127.0.0.1;Port=5432;Database=wysg_dev;Username=postgres;Password=`123qweas"
========================================================
Changes i want:
i want the default value (the value before local setting is saved) of the local connection string to be "Host=127.0.0.1;Port=5432;Database=wysg_dev;Username=postgres"
i want the default value (the value before local setting is saved) of the snowstorm base URL to be "http://127.0.0.1:8080/"
========================================================
Current state:
currently the API path is hard coded as "http://localhost:5205/" or other string like that.
========================================================
Changes i want:
i want to be able to set and save the path of API in the settings window -> network tab, the default value (the value before local setting is saved) being "http://127.0.0.1:5205/"
========================================================
Current state:
the application shouldn't work if the API Base URL is wrong, but when i changed the API Base URL from "http://127.0.0.1:5205/" to "http://127.0.0.1:520/" the app is still working well. i suspect that the API base URL in the network tab is just a skeleton and is not used.
========================================================
Changes i want:
I want the whole app's code related to the radium API to use the API Base URL in the settings window -> network tab, without any fallback.
========================================================